+++ eth_defi/csigma/__init__.py
@@ -1,0 +1,1 @@
+"""cSigma Finance protocol integration."""